submissions/AnikaSystems_Submission/AnikaSystems Code and Data/AnikaSystems Source Code/requirements.txt
flask==1.1.1
waitress==1.4.4
termcolor==1.1.0
requests==2.22.0
pymongo==3.10.1
pdf2image==1.8.0
pytesseract==0.3.4
pywin32==224
python-docx==0.8.10
torch==1.6.0+cu101
pandas==0.25.3
transformers==3.0.2